requirements.txt
pocket-api
requests
